+++ python_modules/libraries/dagster-dg-cli/dagster_dg_cli/cli/api/asset/__init__.py
@@ -1,0 +1,5 @@
+"""Asset API subpackage."""
+
+from dagster_dg_cli.cli.api.asset.commands import asset_group
+
+__all__ = ["asset_group"]
