+++ python_modules/libraries/dagster-dg-cli/dagster_dg_cli/cli/api/asset/display_models.py
@@ -1,6 +1,6 @@
-"""Intermediate display models for consistent formatting across output formats.
+"""Asset display models for consistent formatting across output formats.
 
-These models represent the normalized view of data that should be displayed,
+These models represent the normalized view of asset data that should be displayed,
 ensuring consistency across table, JSON, and markdown formats.
 """
 
@@ -11,7 +11,6 @@ from dagster._record import record
 
 if TYPE_CHECKING:
     from dagster_dg_cli.api_layer.schemas.asset import DgApiAsset, DgApiAssetList
-    from dagster_dg_cli.api_layer.schemas.deployment import Deployment, DeploymentList
 
 
 @record
@@ -63,23 +62,6 @@ class AssetListDisplay:
     items: list[AssetListItemDisplay]
     cursor: Optional[str]
     has_more: bool
-
-
-@record
-class DeploymentDisplay:
-    """Normalized display representation of a deployment."""
-
-    name: str
-    id: str
-    type: str
-
-
-@record
-class DeploymentListDisplay:
-    """Normalized display representation of a deployment list."""
-
-    items: list[DeploymentDisplay]
-    total: int
 
 
 def _extract_metadata_value(entry: dict) -> str:
@@ -230,33 +212,4 @@ def asset_list_to_display(assets: "DgApiAssetList") -> AssetListDisplay:
         items=items,
         cursor=assets.cursor,
         has_more=assets.has_more,
-    )
-
-
-def deployment_to_display(deployment: "Deployment") -> DeploymentDisplay:
-    """Convert API deployment model to display model."""
-    from dagster_dg_cli.api_layer.schemas.deployment import Deployment
-
-    check.inst_param(deployment, "deployment", Deployment)
-
-    return DeploymentDisplay(
-        name=deployment.name,
-        id=str(deployment.id),
-        type=deployment.type.value,
-    )
-
-
-def deployment_list_to_display(deployments: "DeploymentList") -> DeploymentListDisplay:
-    """Convert API deployment list model to display model."""
-    from dagster_dg_cli.api_layer.schemas.deployment import DeploymentList
-
-    check.inst_param(deployments, "deployments", DeploymentList)
-
-    items = []
-    for deployment in deployments.items:
-        items.append(deployment_to_display(deployment))
-
-    return DeploymentListDisplay(
-        items=items,
-        total=deployments.total,
     )
