+++ python_modules/libraries/dagster-dg-cli/dagster_dg_cli/cli/api/deployment/__init__.py
@@ -1,0 +1,5 @@
+"""Deployment API subpackage."""
+
+from dagster_dg_cli.cli.api.deployment.commands import deployment_group
+
+__all__ = ["deployment_group"]
