+++ python_modules/libraries/dagster-dg-cli/dagster_dg_cli/cli/api/deployment/commands.py
@@ -8,9 +8,9 @@ from dagster_dg_core.utils.telemetry import cli_telemetry_wrapper
 from dagster_shared.plus.config import DagsterPlusCliConfig
 from dagster_shared.plus.config_utils import dg_api_options
 
-# Lazy import to avoid loading pydantic at CLI startup
 from dagster_dg_cli.cli.api.client import create_dg_api_graphql_client
-from dagster_dg_cli.cli.api.formatters import format_deployments
+from dagster_dg_cli.cli.api.deployment.formatters import format_deployments
+from dagster_dg_cli.cli.api.shared import determine_output_format
 
 
 @click.command(name="list", cls=DgClickCommand, unlaunched=True)
@@ -38,17 +38,7 @@ def list_deployments_command(
     view_graphql: bool,
 ) -> None:
     """List all deployments in the organization."""
-    # Validate that only one output format is specified
-    if output_json and output_md:
-        raise click.UsageError("Cannot specify both --json and --md flags")
-
-    # Determine output format
-    if output_json:
-        output_format = "json"
-    elif output_md:
-        output_format = "markdown"
-    else:
-        output_format = "table"
+    output_format = determine_output_format(output_json, output_md)
 
     config = DagsterPlusCliConfig.create_for_organization(
         organization=organization,
